refactor(tests): #58 Tier 1.2 — split scheduler test file by feature
tests/test_calendar_refresh_scheduler.py (2,635 LOC, 77 tests) →
tests/calendar/scheduler/{refresh,baseline,breaker,isolation,burst}.py.

Banner-driven split following the file's existing section dividers, plus
peeling the three ConnectorState-primitive tests out of the breaker
section into baseline.py per the issue's plan:

  test_refresh.py     24 tests — refresh_all_schedules core invariants +
                                 service-op wiring + sweep_value_side
                                 (789 LOC)
  test_baseline.py     3 tests — mark_failure / mark_success state primitives
                                 (88 LOC)
  test_breaker.py     15 tests — circuit-breaker driver tests
                                 (563 LOC)
  test_isolation.py   20 tests — daily request-budget tracking (BLS cap)
                                 (661 LOC)
  test_burst.py       15 tests — schedule-aware burst (issue #37)
                                 (616 LOC)

Each split file imports only the symbols its tests reference, redeclares
the `store` fixture and `_FakeSummary` dataclass it uses, and inlines
`_insert_pending_event` into the burst file (its sole caller). Test
bodies are byte-equivalent to the original.

Verification: 77 tests pass on branch (same count as master).
Codex review confirmed the split preserves test bodies and discovery.
